Introduce file_operations structure and replace fs_ls with fs_show_files
Introduce file_operations structure to implement a number of callback
functions. With the help of the structure, the function register_fs just
needs a pointer to file_operations structure. We do not need to pass each
argument of callback functions to the function register_fs.

This commit also replaces the function fs_ls with fs_show_files.